refactor: dao 인터페이스 분리
- 변경할 때 새로운 구현체로 변경하기 쉽도록 분리합니다.
