+++ src/main/java/janggi/GameMode.java
@@ -8,6 +8,10 @@ import janggi.data.dao.BoardDao;
 import janggi.data.dao.CampDao;
 import janggi.data.dao.PieceDao;
 import janggi.data.dao.PieceSymbolDao;
+import janggi.data.dao.mysql.MySqlBoardDao;
+import janggi.data.dao.mysql.MySqlCampDao;
+import janggi.data.dao.mysql.MySqlPieceDao;
+import janggi.data.dao.mysql.MySqlPieceSymbolDao;
 import janggi.view.View;
 
 enum GameMode {
@@ -25,13 +29,12 @@ enum GameMode {
     }
 
     public Controller createController() {
+        CampDao campDao = new MySqlCampDao();
+        PieceSymbolDao pieceSymbolDao = new MySqlPieceSymbolDao();
+        BoardDao boardDao = new MySqlBoardDao(campDao);
+        PieceDao pieceDao = new MySqlPieceDao(boardDao, pieceSymbolDao);
         return switch (this) {
-            case ONLINE -> new OnlineController(
-                    new View(),
-                    new CampDao(),
-                    new PieceSymbolDao(),
-                    new BoardDao(),
-                    new PieceDao());
+            case ONLINE -> new OnlineController(new View(), campDao, pieceSymbolDao, boardDao, pieceDao);
             case OFFLINE -> new OfflineController(new View());
         };
     }
